requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.11.6
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==5.0.1
asyncio==3.4.3
attrs==24.2.0
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
decorator==4.4.2
distro==1.9.0
elevenlabs==1.12.1
exceptiongroup==1.2.2
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
jiter==0.7.1
leonardo_api==0.0.11
moviepy==1.0.3
multidict==6.1.0
numpy==2.1.3
openai==1.54.5
opencv-python==4.10.0.84
pillow==10.4.0
proglog==0.1.10
propcache==0.2.0
pydantic==2.9.2
pydantic_core==2.23.4
pysrt==1.1.2
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
sniffio==1.3.1
tqdm==4.67.0
typing_extensions==4.12.2
urllib3==2.2.3
websockets==14.1
yarl==1.17.2
yt-dlp==2024.11.18
